feat: Migrate account and management views to users app 🎉✨ Refactored the account and management views to the users app for better organization and maintainability. 🛠️🔧 Updated URL routing to reflect these changes, ensuring proper access to user account functionalities. 🚀💻 Removed old templates and created new ones in the accounts directory for a cleaner structure. 🗂️📄 Enhanced test cases to verify the correct template usage for the account page. ✅📝
