+++ app/urls/web.py
@@ -21,7 +21,5 @@ urlpatterns = [
         views.sample_types_pivot,
         name="sample_types_pivot",
     ),
-    path("account/", views.account, name="account"),
     path("data_export/", views.data_export, name="data_export"),
-    path("management/", views.management, name="management"),
 ]
