+++ app/views/sample_views.py
@@ -167,29 +167,6 @@ def used_samples_archive_all(request):
 
 
 @login_required(login_url="/login/")
-def account(request):
-    # User account page showing last 20 recently accessed samples
-    sample_list = (
-        Sample.objects.filter(last_modified_by=request.user.email)
-        .select_related("study_id")
-        .order_by("-last_modified")[:20]
-    )
-    context = {"sample_list": sample_list}
-    return render(request, "samples/account.html", context)
-
-
-@login_required(login_url="/login/")
-def management(request):
-    users = User.objects.all()
-    user_email_list = []
-    for user in users:
-        user_email_list.append(user.email)
-    user_email_list = ";".join(user_email_list)
-    context = {"user_email_list": user_email_list}
-    return render(request, "samples/management.html", context)
-
-
-@login_required(login_url="/login/")
 def sample_add(request):
     # Add new sample page
     if request.method == "POST":
